requirements.txt
flask==3.0.0
ollama==0.1.7
pytesseract==0.3.10
pillow==10.0.0
numpy==1.24.0
reportlab==4.0.8
